refactor: extract remaining service domain implementations
