+++ scripts/release_artifact_allowlist.txt
@@ -225,12 +225,16 @@ src/diptrace_mcp/services/evidence.py
 src/diptrace_mcp/services/exports.py
 src/diptrace_mcp/services/external_jobs.py
 src/diptrace_mcp/services/jobs.py
+src/diptrace_mcp/services/live_sessions.py
 src/diptrace_mcp/services/placement.py
 src/diptrace_mcp/services/review.py
 src/diptrace_mcp/services/routing.py
 src/diptrace_mcp/services/scaffolding.py
+src/diptrace_mcp/services/semantic_engine.py
 src/diptrace_mcp/services/semantic_operations.py
+src/diptrace_mcp/services/synchronization.py
 src/diptrace_mcp/services/transactions.py
+src/diptrace_mcp/services/xml_writes.py
 src/diptrace_mcp/sessions.py
 src/diptrace_mcp/silkscreen.py
 src/diptrace_mcp/spatial.py
